backend/requirements.txt
asgiref==3.9.1
cffi==1.17.1
channels==4.0.0
channels-redis==4.1.0
cryptography==41.0.7
Django==5.2.5
django-cors-headers==4.3.1
django-ratelimit==4.1.0
django-redis==5.4.0
djangorestframework==3.14.0
msgpack==1.1.1
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.22
pypng==0.20220715.0
python-decouple==3.8
pytz==2025.2
qrcode==7.4.2
redis==5.0.1
sqlparse==0.5.3
typing_extensions==4.15.0
uuid==1.30
websockets==12.0
